Animate web server, websocket handler, gen_server
A web page to hold a canvas and a websocket, a gen_server to handle
animation commands and animating and a Cowboy Websocket handler to
connect the two.

Currently only features start and stop buttons on the web page that
start and stop a constant stream of randomly generated points from the
animation gen_server.

Issues:
- hard coded max X and max Y coordinates
- No animation commands, just random coordinates
- No animation, just static canvas images
- Only one animation server (should have one per websocket connection)