Move command parsing, reconnect animator
Moved the height and width command parsing to the websocket handler from
the animator; divided into two separate calls. Makes more sense to have
the Websocket handler do the parsing than the animator. Also made more
sense to set the height and width with different calls.

If the supervisor restarts a crashed animator the animator will notify
the websocket handler of the new animator PID.

Issues:
  - No animation, just squares
  - Restarted animator has no height and width

I think I'm going to start a second process for the animator to store
state from the web page.