Animate heat map with linear falloff
Animate a heat map for each boid in a separate canvas. The boids don't
yet consider the heat map for movement; I just wanted to get the
animation working first.

I moved the shape drawing to shape.erl and moved the JSON encoding to
buffer.erl and heatmap.erl just before the shapes are sent to the
Websocket. I could probably just have the Websocket handler do the JSON
encoding, since it's right at the point where JSON is needed: everywhere
else it would be easier to work with Erlang data.