Rudimentary flocking
I've got the boids checking the heat map for the best surrounding cells
to go to. They'll look for the warmest spot near them that isn't too
warm. If all the spots are too warm, they'll look for the coolest. If
none of the spots are warm they'll pick a random direction.

So far the flocking just turns out to be all the boids moving in one
direction. Success, but not exciting. I need to work out either keeping
the boids from going outside the canvas or teleporting them to the other
side so they don't disappear off the edge.