Switch Buffer ETS to Dict, cleanup, speedup
Duh. I switched to a dict in the buffer so I could update the stored
shapes by key but I could have just stuck with an ETS table and done the
same. Oh Well. Anywho, by updating existing shapes and tracking them by
boid PID I don't "lose" boids whose PID process don't manage to get them
into the buffer in the current cycle. Previously, if a boid process
didn't get a shape into the buffer before the buffer sent the shapes to
the canvas then the canvas is blanked out and the missed boids are
absent on the redraw. Now the buffer will keep sending old shapes to the
canvas.

I also sped up the cycle times to get a clearer idea of the boid
behaviour. Currently the boids get stuck in a "rut" where they're too
far away so they move a little closer; now they're too close so they
move back, and so on, vibrating back and forth in a static pattern.